+++ src/lib/converter/utils/nodes.ts
@@ -4,11 +4,7 @@ export function isNamedNode(node: ts.Node): node is ts.Node & {
     name: ts.Identifier | ts.PrivateIdentifier | ts.ComputedPropertyName;
 } {
     const name: ts.Node | undefined = (node as any).name;
-    return (
-        !!name &&
-        (ts.isIdentifierOrPrivateIdentifier(name) ||
-            ts.isComputedPropertyName(name))
-    );
+    return !!name && (ts.isMemberName(name) || ts.isComputedPropertyName(name));
 }
 
 export function getHeritageTypes(